pyproject.toml
[project]
name = "quant-news"
version = "0.1.0"
description = "Quantitative stock tracking app with news analysis"
requires-python = ">=3.10"
dependencies = [
    "dash>=2.15.0",
    "dash-bootstrap-components>=1.5.0",
    "plotly>=5.18.0",
    "yfinance>=0.2.36",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "duckdb>=0.10.0",
    "pyarrow>=15.0.0",
    "ta>=0.11.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | __pycache__
  | dist
  | build
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["config", "services", "layouts", "callbacks", "utils"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = [".git", ".mypy_cache", ".pytest_cache", ".venv", "__pycache__"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["tests/", ".venv/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", ".mypy_cache", ".pytest_cache", ".venv", "__pycache__", "dist", "build"]
